almost finishes endpoint infrastructure

+++ src/main/java/MoEzwawi/GUIDANOMADI_BACKEND/services/FavouritesService.java
@@ -12,7 +12,6 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 
-import java.util.List;
 import java.util.Optional;
 
 @Service
@@ -22,19 +21,27 @@ public class FavouritesService {
     public Favourite findById(Long id){
         return this.favouritesRepository.findById(id).orElseThrow(()-> new NotFoundException(id));
     }
+    public Optional<Favourite> findByUserAndProperty(User currentUser, Property property){
+        return this.favouritesRepository.findByUserAndProperty(currentUser, property);
+    }
+    public boolean isItFavourite(User currentUser, Property property){
+        return this.findByUserAndProperty(currentUser, property).isPresent();
+    }
     public Favourite addToFavourites(User currentUser, Property property){
         return this.favouritesRepository.save(new Favourite(currentUser,property));
-    }
-    public void removeFromFavourites(Long id){
-        Favourite found = this.findById(id);
-        this.favouritesRepository.delete(found);
     }
     public Page<Favourite> getMyFavourites(int page, int size, String orderBy, User currentUser){
         Pageable pageable = PageRequest.of(page, size, Sort.by(orderBy));
         return this.favouritesRepository.findByUser(currentUser, pageable);
     }
-    public boolean isItFavourite(User currentUser, Property property){
-        Optional<Favourite> optionalFavourite = this.favouritesRepository.findByUserAndProperty(currentUser, property);
-        return optionalFavourite.isPresent();
+    public void removeFromFavourites(Long id){
+        Favourite found = this.findById(id);
+        this.favouritesRepository.delete(found);
+    }
+    public void removeFromFavourites(User currentUser, Property property){
+        if (this.findByUserAndProperty(currentUser, property).isPresent()){
+            Favourite found = this.findByUserAndProperty(currentUser, property).get();
+            this.favouritesRepository.delete(found);
+        }
     }
 }
